Add support for ctr2tcon
I don't think this should actually be supported by `rock_properties` itself; the conversion from CTR to conductivity is substantial enough to warrant its own tool and output files, but I'm including it like this for now in the name of backwards compatibility.

The Matlab version uses and optimisation function to solve this problem, but that isn't strictly necessary for polynomial CTR profiles (which is all we have) - so I've opted to remove the optimiser and do a simple derivative instead. This simplifies the code, but does yield slightly different (1-2%) results. I consider this acceptable, and will inspect all of the JdF fixture files to ensure there are no deviations beyond this.